% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/too-many-cells.R
\name{importResults}
\alias{importResults}
\title{Import some too-many-cells make-tree outputs into a data frame.}
\usage{
importResults(dir = "out")
}
\arguments{
\item{dir}{The output directory of a too-many-cells run.}
}
\value{
A list of each output. Reads the following files, see
  \url{https://gregoryschwartz.github.io/too-many-cells/} for more details:
  "dendrogram.svg", "clumpiness.pdf", "projection.pdf",
  "label_projection.pdf", "clumpiness.csv", "cluster_info.csv",
  "node_info.csv", and "cluster_diversity.csv".
}
\description{
This function will import some of the files resulting from a too-many-cells
make-tree run into R as data frames. Does not include cluster list. Look at
the main tooManyCells function for the cluster list.
}
\examples{
\dontrun{
res = importResults("out")
plot(res$treePlot, axes = FALSE)
res$clumpiness
plot(res$clumpinessPlot, axes = FALSE)
}
}
