% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/too-many-cells.R
\name{tooManyCells}
\alias{tooManyCells}
\title{Execute too-many-cells.}
\usage{
tooManyCells(mat, args = c("make-tree"), labels = NULL,
  output = "out", prior = NULL)
}
\arguments{
\item{mat}{The input Matrix with gene row names and
cell barcode column names.}

\item{args}{The arguments to give to the command line program. See
\url{https://gregoryschwartz.github.io/too-many-cells/} for more information.
Defaults to "make-tree".}

\item{labels}{The input labels data frame with item (cell barcodes) and label
(whatever labels you want to give them, such as tissue of origin, celltype,
etc.) columns. Optional.}

\item{output}{The output folder for the too-many-cells process. Defaults to
"out".}

\item{prior}{The location of the tree that was already made (previous
too-many-cells output) so quick visual or pruning changes can be made
without remaking the tree (can potentially save hours).}
}
\value{
A list of each output, including the stdout. Reads the
  following files, see \url{https://gregoryschwartz.github.io/too-many-cells/} for
  more details: "dendrogram.svg", "clumpiness.pdf", "projection.pdf",
  "label_projection.pdf", "clumpiness.csv", "cluster_info.csv",
  "node_info.csv", and "cluster_diversity.csv".
}
\description{
This function will run too-many-cells on a Matrix.
  Requires too-many-cells to be installed (follow instructions at
  \url{https://gregoryschwartz.github.io/too-many-cells/} ).
}
\examples{
\dontrun{
tooManyCells(mat, args = c("make-tree", "--smart-cutoff", "4", "--min-size", "1"))
plot(res$treePlot, axes = FALSE)
res$stdout
res$nodeInfo
plot(res$clumpinessPlot, axes = FALSE)
}
}
