% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/too-many-cells.R
\name{tryFunc}
\alias{tryFunc}
\title{Different error for importing data.}
\usage{
tryFunc(f, file)
}
\arguments{
\item{f}{The function to use.}

\item{file}{The input file to be read.}
}
\value{
The imported data frame or NULL if an error occurred.
}
\description{
This function will fail gracefully instead of stopping the program with an
error.
}
\examples{
\dontrun{
x = tryFunc(read.csv, "input.csv")
}
}
