% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/too-many-cells.R
\name{writeMatrix}
\alias{writeMatrix}
\title{Write a Matrix to a folder.}
\usage{
writeMatrix(mat, labels = NULL)
}
\arguments{
\item{mat}{The input Matrix with gene row names and cell barcode column
names.}

\item{labels}{The input labels data frame with item (cell barcodes) and label
(whatever labels you want to give them, such as tissue of origin, celltype,
etc.) columns. Optional.}
}
\value{
None
}
\description{
This function will write a Matrix from the Matrix library to a temporary
directory containing matrix.mtx, genes.tsv, barcodes.tsv, and optionally a
labels.csv file.
}
\examples{
\dontrun{
library(Matrix)
df = read.csv("mat.csv", row.names = 1, header = TRUE)
mat = Matrix(as.matrix(in), sparse = TRUE)
writeMatrix(mat)
}
}
