% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TooManyCellsR.R
\name{tryFunc}
\alias{tryFunc}
\title{Different error for importing data.}
\usage{
tryFunc(f, file)
}
\arguments{
\item{f}{The function to use.}

\item{file}{The input file to be read.}
}
\value{
The imported data frame or NULL if an error occurred.
}
\description{
This function will fail gracefully instead of stopping the program with an
error.
}
\examples{
input <- system.file("extdata", "mat.csv", package="TooManyCellsR")
fail = tryFunc(read.csv, "fail.csv")
fail
success = tryFunc(read.csv, input)
success
}
